+++ docs/ReleaseNotes.md
@@ -1,3 +1,4 @@
 # 0.1.0.0
 
 - Basic Layout
+- local storage implemented